requirements.txt
alembic==1.13.1
Babel==2.14.0
black==23.12.1
boto3==1.28.54
cryptography==42.0.2
email-validator==2.1.0.post1
factory-boy==3.3.0
fastapi==0.108.0
flake8==6.1.0
httpx==0.26.0
isort==5.13.2
mypy==1.8.0
pip==24.0
pydantic-settings==2.1.0
PyMySQL==1.1.0
pytest==7.4.4
pytz==2024.1
sqlalchemy-utils==0.41.1
sqlmodel==0.0.18
setuptools==65.5.1
uvicorn==0.25.0
wheel==0.42.0
uvicorn[standard]
